requirements.txt
asgiref==3.3.4
automium==0.2.6
automium-web==0.1.1
click==8.0.1
colorama==0.4.4
convertdate==2.3.2
Django==3.2.4
django-crispy-forms==1.11.2
django-parsley==0.7
djangorestframework==3.12.4
djangorestframework-datatables==0.6.0
Flask==2.0.1
hijri-converter==2.1.2
holidays==0.11.1
itsdangerous==2.0.1
Jinja2==3.0.1
korean-lunar-calendar==0.2.1
legacy==0.1.6
MarkupSafe==2.0.1
mysql-connector==2.2.9
mysqlclient==2.0.3
Pillow==8.2.0
PyMeeus==0.5.11
python-dateutil==2.8.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
Werkzeug==2.0.1
